# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wossolaty API Documentation",
        "description": "This is a wossolaty API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "haidershamil@gmail"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/receipts": {
            "get": {
                "tags": [
                    "Receipt"
                ],
                "summary": "Returns API response of all Receipt",
                "description": "A sample test of the API",
                "operationId": "getReceiptList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Receipt"
                ],
                "summary": "Store new Receipt",
                "description": "Returns Receipt data",
                "operationId": "storeReceipt",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/receipts/{id}": {
            "get": {
                "tags": [
                    "Receipt"
                ],
                "summary": "Get Receipt information",
                "description": "Returns Receipt data",
                "operationId": "getReceiptById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Receipt id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Receipt"
                ],
                "summary": "Update existing Receipt",
                "description": "Returns updated Receipt data",
                "operationId": "updateReceipt",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Receipt id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Receipt"
                ],
                "summary": "Delete existing Receipt",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteReceipt",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Receipt id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/stocks": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Returns API response of all Stock",
                "description": "A sample test of the API",
                "operationId": "getStockList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Store new Stock",
                "description": "Returns Stock data",
                "operationId": "storeStock",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/stocks/{id}": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Get Stock information",
                "description": "Returns Stock data",
                "operationId": "getStockById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Stock id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Stock"
                ],
                "summary": "Update existing Stock",
                "description": "Returns updated Stock data",
                "operationId": "updateStock",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Stock id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Stock"
                ],
                "summary": "Delete existing Stock",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteStock",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Stock id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
...
}